requirements.txt
flask
pandas
reportlab
python-barcode
pillow
openpyxl
